# optimize parsing common timestamps